fix migrations - method: 'down'

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -30,6 +30,8 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::drop('users');
+        if (Schema::hasTable('users')) {
+            Schema::drop('users');
+        }
     }
 }
