+++ database/migrations/2018_01_09_204250_create_teams_table.php
@@ -27,6 +27,8 @@ class CreateTeamsTable extends Migration
      */
     public function down()
     {
-        Schema::drop('teams');
+        if (Schema::hasTable('teams')) {
+            Schema::drop('teams');
+        }
     }
 }
